requirements.txt
# Core ML
scikit-learn>=1.3.0
xgboost>=2.0.0
joblib>=1.3.0

# Data Processing
pandas>=2.1.0
numpy>=1.24.0

# Visualization
matplotlib>=3.8.0
seaborn>=0.13.0

# Wazuh Integration
requests>=2.31.0
schedule>=1.2.0
pyyaml>=6.0.0

# Notebooks
jupyter>=1.0.0
notebook>=7.0.0

# Utilities
tqdm>=4.66.0
